친구 기능 최종 수정

+++ daerkoob_back/src/main/java/com/project/daerkoob/controller/ReviewController.java
@@ -54,7 +54,7 @@ public class ReviewController {
     }
 
     @GetMapping("register/{userId}/{isbn}/{reviewContent}/{score}") //guide line , 이제 그냥 isbn 넘겨주시면 가능합니다.
-    public boolean getRegister(@PathVariable Long userId, @PathVariable String isbn, @PathVariable String reviewContent , @PathVariable Long score) throws Exception{
+    public boolean getRegister(@PathVariable Long userId, @PathVariable String isbn, @PathVariable String reviewContent , @PathVariable Double score) throws Exception{
         bookService.save(bookService.createBook(isbn));
         Optional<Book> book = bookService.findBook(isbn);
         reviewService.save(reviewService.createDto(userId, book.get().getId() ,  score , reviewContent));
@@ -63,7 +63,7 @@ public class ReviewController {
 
 
     @PostMapping("register")
-    public boolean register(Long userId, String isbn , String reviewContent , Long score) throws Exception{
+    public boolean register(Long userId, String isbn , String reviewContent , Double score) throws Exception{
         bookService.save(bookService.createBook(isbn));
         Optional<Book> book = bookService.findBook(isbn);
         reviewService.save(reviewService.createDto(userId, book.get().getId() ,  score , reviewContent));
